refactor: common 패턴 적용
