--- .env.example
@@ -1,22 +1,0 @@
-DB_TYPE=mysql
-
-DB_USER=
-DB_PASS=
-DB_HOST=
-DB_NAME=
-DB_PORT=
-
-SERVER_HOST=
-SERVER_PORT=
-
-JWT_SECRET=
-
-KAKAO_REST_API_KEY=
-
-SWAGGER_USER=
-SWAGGER_PASSWORD=
-
-AWS_S3_ACCESS_KEY=
-AWS_S3_SECRET_KEY=
-AWS_S3_REGION=
-AWS_S3_BUCKET_NAME=
