+++ src/app.module.ts
@@ -7,23 +7,23 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProfilesModule } from './profiles/profiles.module';
 import { PersonaModule } from './persona/persona.module';
 import { AuthModule } from './auth/auth.module';
-import { JwtMiddleware } from './_middleware/jwt.middleware';
-import { Categories } from './_entities/Categories';
-import { FeedCategoryMapping } from './_entities/FeedCategoryMapping';
-import { FeedHashTagMapping } from './_entities/FeedHashTagMapping';
-import { FeedImgs } from './_entities/FeedImgs';
-import { Feeds } from './_entities/Feeds';
-import { FollowFromTo } from './_entities/FollowFromTo';
-import { HashTags } from './_entities/HashTags';
-import { Likes } from './_entities/Likes';
-import { Notice } from './_entities/Notice';
-import { Persona } from './_entities/Persona';
-import { ProfileHashTagMapping } from './_entities/ProfileHashTagMapping';
-import { Profiles } from './_entities/Profiles';
-import { QuestionContent } from './_entities/QuestionContent';
-import { Questions } from './_entities/Questions';
-import { Users } from './_entities/Users';
-import { Environment } from './_utilities/constants';
+import { JwtMiddleware } from './common/middleware/jwt.middleware';
+import { Categories } from './common/entities/Categories';
+import { FeedCategoryMapping } from './common/entities/FeedCategoryMapping';
+import { FeedHashTagMapping } from './common/entities/FeedHashTagMapping';
+import { FeedImgs } from './common/entities/FeedImgs';
+import { Feeds } from './common/entities/Feeds';
+import { FollowFromTo } from './common/entities/FollowFromTo';
+import { HashTags } from './common/entities/HashTags';
+import { Likes } from './common/entities/Likes';
+import { Notice } from './common/entities/Notice';
+import { Persona } from './common/entities/Persona';
+import { ProfileHashTagMapping } from './common/entities/ProfileHashTagMapping';
+import { Profiles } from './common/entities/Profiles';
+import { QuestionContent } from './common/entities/QuestionContent';
+import { Questions } from './common/entities/Questions';
+import { Users } from './common/entities/Users';
+import { Environment } from './common/utils/constants';
 import { LikesModule } from './likes/likes.module';
 import { FollowingModule } from './following/following.module';
 import { FeedsModule } from './feeds/feeds.module';
