+++ src/auth/auth.module.ts
@@ -3,11 +3,11 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserService } from './user.service';
-import { Users } from '../_entities/Users';
+import { Users } from '../common/entities/Users';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { JwtStrategy } from './security/jwt.strategy';
-import { jwtConfig } from '../_config/jwt.config';
+import { jwtConfig } from '../../config/jwt.config';
 import { KakaoService } from './kakao/kakao.service';
 import { GoogleStrategy } from './security/google.strategy';
 import { GoogleService } from './google/google.service';
