+++ src/common/entities/FeedCategoryMapping.ts
