+++ src/common/entities/Persona.ts
