+++ src/feeds/feeds.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { LikesRepository } from '../likes/likes.repository';
-import { Feeds } from '../_entities/Feeds';
-import { Likes } from '../_entities/Likes';
+import { Feeds } from '../common/entities/Feeds';
+import { Likes } from '../common/entities/Likes';
 import { FeedsController } from './feeds.controller';
 import { FeedRepository } from './feeds.repository';
 import { FeedsService } from './feeds.service';
