+++ src/following/following.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FollowFromTo } from '../_entities/FollowFromTo';
+import { FollowFromTo } from '../common/entities/FollowFromTo';
 import { FollowingController } from './following.controller';
 import { FollowingRepository } from './following.repository';
 import { FollowingService } from './following.service';
