+++ src/main.ts
@@ -3,7 +3,7 @@ import { AppModule } from './app.module';
 import * as expressBasicAuth from 'express-basic-auth';
 import { SwaggerModule, DocumentBuilder, OpenAPIObject } from '@nestjs/swagger';
 import * as cookieParser from 'cookie-parser';
-import { BaseAPIDocument } from './_config/swagger.document.config';
+import { BaseAPIDocument } from '../config/swagger.document.config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
