+++ src/persona/persona.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Persona } from '../_entities/Persona';
+import { Persona } from '../common/entities/Persona';
 import { PersonaController } from './persona.controller';
 import { PersonaRepository } from './persona.repository';
 import { PersonaService } from './persona.service';
