+++ src/profiles/profiles.controller.ts
@@ -1,5 +1,5 @@
 import { AwsService } from './../aws/aws.service';
-import { multerOptions } from './../_utilities/multer.option';
+import { multerOptions } from '../common/utils/multer.option';
 import {
   Body,
   Controller,
@@ -24,8 +24,8 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { JWTAuthGuard } from '../auth/security/auth.guard.jwt';
-import baseResponse from '../_utilities/baseResponseStatus';
-import { errResponse, sucResponse } from '../_utilities/response';
+import baseResponse from '../common/utils/baseResponseStatus';
+import { errResponse, sucResponse } from '../common/utils/response';
 import { CreateProfileDto } from './dto/createProfile.dto';
 import { EditProfileDto } from './dto/editProfile.dto';
 import { ProfileModelExample } from './dto/profile.model';
