+++ src/profiles/profiles.module.ts
@@ -3,8 +3,8 @@ import { Module } from '@nestjs/common';
 import { MulterModule } from '@nestjs/platform-express';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PersonaRepository } from '../persona/persona.repository';
-import { Persona } from '../_entities/Persona';
-import { Profiles } from '../_entities/Profiles';
+import { Persona } from '../common/entities/Persona';
+import { Profiles } from '../common/entities/Profiles';
 import { ProfilesController } from './profiles.controller';
 import { ProfilesRepository } from './profiles.repository';
 import { ProfilesService } from './profiles.service';
