+++ src/profiles/profiles.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { PersonaRepository } from '../persona/persona.repository';
 import { SaveProfileDto } from './dto/saveProfile.dto';
-import baseResponse from '../_utilities/baseResponseStatus';
-import { errResponse, sucResponse } from '../_utilities/response';
+import baseResponse from '../common/utils/baseResponseStatus';
+import { errResponse, sucResponse } from '../common/utils/response';
 import { CreateProfileDto } from './dto/createProfile.dto';
 import { ProfilesRepository } from './profiles.repository';
 import { EditProfileDto } from './dto/editProfile.dto';
